Move forwarding logic and template into a separate component
