+++ src/components/MessagesList/MessagesGroup/Message/Message.vue
@@ -164,8 +164,9 @@ the main body of the message as well as a quote.
 								{{ t('spreed', 'Go to file') }}
 							</ActionLink>
 							<ActionButton
+								v-if="!isFileShare"
 								:close-after-click="true"
-								@click.stop="handleForwardMessage">
+								@click.stop="showForwarder = true">
 								<Share
 									slot="icon"
 									:size="16"
@@ -204,32 +205,9 @@ the main body of the message as well as a quote.
 				<span>{{ t('spreed', 'Unread messages') }}</span>
 			</div>
 		</div>
-
-		<!-- Message forwarding -->
-		<RoomSelector
-			v-if="showRoomSelector"
-			container="#content-vue"
-			:show-postable-only="true"
-			:dialog-title="dialogTitle"
-			:dialog-subtitle="dialogSubtitle"
-			@select="setSelectedConversation"
-			@close="showRoomSelector=false" />
-		<Modal v-if="showForwardedConfirmation"
-			@close="showForwardedConfirmation = false">
-			<EmptyContent icon="icon-checkmark" class="forwarded-confirmation__emptycontent">
-				<template #desc>
-					{{ t('spreed', 'The message has been forwarded to {selectedConversation}', { selectedConversation }) }}
-				</template>
-			</EmptyContent>
-			<div class="forwarded-confirmation__navigation">
-				<button @click="showForwardedConfirmation=false">
-					{{ t('spreed', 'Dismiss') }}
-				</button>
-				<button class="primary" @click="openConversation(room)">
-					{{ t('spreed', 'Go to conversation') }}
-				</button>
-			</div>
-		</Modal>
+		<Forwarder v-if="showForwarder"
+			:message-object="messageObject"
+			@close="showForwarder = false" />
 	</li>
 </template>
 
@@ -267,9 +245,7 @@ import {
 import { generateUrl } from '@nextcloud/router'
 import Location from './MessagePart/Location'
 import Contact from './MessagePart/Contact.vue'
-import RoomSelector from '../../../../views/RoomSelector.vue'
-import EmptyContent from '@nextcloud/vue/dist/Components/EmptyContent'
-import Modal from '@nextcloud/vue/dist/Components/Modal'
+import Forwarder from './MessagePart/Forwarder'
 
 export default {
 	name: 'Message',
@@ -292,9 +268,7 @@ export default {
 		Reload,
 		Share,
 		ActionSeparator,
-		RoomSelector,
-		EmptyContent,
-		Modal,
+		Forwarder,
 	},
 
 	mixins: [
@@ -441,9 +415,8 @@ export default {
 			isDeleting: false,
 			// whether the message was seen, only used if this was marked as last read message
 			seen: false,
-			showRoomSelector: false,
-			selectedConversation: null,
-			showForwardedConfirmation: false,
+			// Shows/hides the message forwarder component
+			showForwarder: false,
 		}
 	},
 
@@ -688,14 +661,6 @@ export default {
 				apiVersion: 'v3',
 			}
 		},
-
-		dialogTitle() {
-			return t('spreed', 'Forward message')
-		},
-
-		dialogSubtitle() {
-			return t('spreed', 'Choose a conversation to which forward the selected message.')
-		},
 	},
 
 	watch: {
@@ -827,25 +792,6 @@ export default {
 
 			// reload conversation to update additional attributes that have computed values
 			await this.$store.dispatch('fetchConversation', { token: this.token })
-		},
-
-		handleForwardMessage() {
-			this.showRoomSelector = true
-		},
-
-		async setSelectedConversation(room) {
-			this.selectedConversation = room
-			this.showRoomSelector = false
-			await this.$store.dispatch('forwardMessage', {
-				token: room,
-				message: this.messageObject,
-			})
-			this.showForwardedConfirmation = true
-
-		},
-		openConversation() {
-			this.$router.push({ name: 'conversation', params: { token: this.selectedConversation } }).catch(err => console.debug(`Error while pushing the new conversation's route: ${err}`))
-			this.showForwardedConfirmation = false
 		},
 	},
 }
@@ -994,22 +940,6 @@ export default {
 
 	&.retry-option {
 		cursor: pointer;
-	}
-}
-
-.forwarded-confirmation {
-	&__emptycontent {
-		width: 280px;
-		text-align: center;
-		margin: 20px !important;
-	}
-	&__navigation {
-		display: flex;
-		justify-content: space-between;
-		padding: 12px;
-		button {
-			height: 44px;
-		}
 	}
 }
 </style>
